requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.8
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
numpy==2.1.2
passlib==1.7.4
psycopg2==2.9.10
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
typing_extensions==4.12.2
annotated-types==0.7.0
anyio==4.8.0
fastapi==0.115.8
greenlet==3.1.1
idna==3.10
numpy==2.1.2
passlib==1.7.4
psycopg2==2.9.10
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
